phidata-exp/requirements.txt
phidata
openai
duckduckgo-search
yfinance
fastapi[standard]
sqlalchemy
packaging
pillow